refactor: change directory to powertrain and add logfile logic to the battery code
