+++ src/powertrain/battery.h
